Algunas modificaciones de los ejercicios

+++ PROGRAMACIÓN_M03_M05/JAVA_DAM_INTELLIJ/src/programacion/RA1/ejerciciosJavaParaProgramar/Ejercicio8.java
@@ -21,7 +21,7 @@ es un 0.*/
 
             if (nota > 0 && nota <= 10)
             {
-                mediaNotas += (double) nota;
+                mediaNotas += nota;
                 contadorNotas++;
             }
             else
